requirements.txt
torch>=2.0
torchvision
pandas
matplotlib
numpy
seaborn
Pillow
scikit-learn
tqdm
requests
torchinfo